--- backend/app/models/api_keys.py
@@ -1,4 +1,0 @@
-# API Keys module - File-backed implementation
-# API keys are now stored in JSON files (data/api_keys/{tenant_id}.json)
-# instead of the database. This file is kept for reference.
-# See: app/services/api_keys.py for the service implementation
